# api/docs/swagger.yaml
basePath: /
definitions:
...
  model.Publication:
    properties:
      authorId:
        type: integer
      authorNick:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      likes:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: API responsible for CRUD and DevBook social network authentication.
  title: DevBook API
  version: "1.0"
paths:
...
  /publications:
    get:
      consumes:
      - application/json
      description: |-
        ListMyAndFollowPublications lists the publications of the logged in user and the publications they follow
        and returns a list containing the posts of the logged in user and the posts of whom he follows or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/model.Publication'
            type: array
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: ListMyAndFollowPublications lists the publications of the logged in
        user and the publications they follow.
      tags:
      - publications
    post:
      consumes:
      - application/json
      description: CreatePublication create a publication and returns the publication
        created or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Enter only the title and content of the publication.
        in: body
        name: publication
        required: true
        schema:
          $ref: '#/definitions/model.Publication'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/model.Publication'
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "422":
          description: 'Error: Unprocessable Entity'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: CreatePublication create a publication.
      tags:
      - publications
  /publications/{publicationId}:
    delete:
      consumes:
      - application/json
      description: |-
        DeletePublicationById deletes a publication by the given ID
        and returns an error if unable to delete the publication.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Publication ID
        in: path
        name: publicationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: DeletePublicationById deletes a publication by the given ID.
      tags:
      - publications
    get:
      consumes:
      - application/json
      description: |-
        FindPublicationById search for a publication by the given ID
        and returns the publication or an error if unable to create the publication.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Publication ID
        in: path
        name: publicationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/model.Publication'
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "404":
          description: 'Error: Not Found'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: FindPublicationById search for a publication by the given ID.
      tags:
      - publications
    put:
      consumes:
      - application/json
      description: |-
        UpdatePublicationById updates a publication by the given ID
        and returns an error if unable to update the publication.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Publication ID
        in: path
        name: publicationId
        required: true
        type: integer
      - description: Enter only the title and content of the publication.
        in: body
        name: publication
        required: true
        schema:
          $ref: '#/definitions/model.Publication'
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "422":
          description: 'Error: Unprocessable Entity'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: UpdatePublicationById updates a publication by the given ID..
      tags:
      - publications
  /publications/{publicationId}/like:
    post:
      consumes:
      - application/json
      description: |-
        LikePublicationById likes a publication by the given ID
        and returns an error if unable to like the publication.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Publication ID
        in: path
        name: publicationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: LikePublicationById likes a publication by the given ID.
      tags:
      - publications
  /publications/{publicationId}/unlike:
    post:
      consumes:
      - application/json
      description: |-
        UnlikePublicationById unlike a publication by the given ID
        and returns an error if unable to unlike the publication.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Publication ID
        in: path
        name: publicationId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: UnlikePublicationById unlike a publication by the given ID.
      tags:
      - publications
  /publications/user/{userId}:
    get:
      consumes:
      - application/json
      description: |-
        ListPublicationsByUserId lists a user's publications
        and returns a list of publications or an error.
      parameters:
      - description: 'Enter the content: Bearer and your access token.'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Success with no content
        "400":
          description: 'Error: Bad Request'
          schema:
            type: string
        "401":
          description: 'Error: Unauthorized'
          schema:
            type: string
        "403":
          description: 'Error: Forbidden'
          schema:
            type: string
        "500":
          description: 'Error: Internal Server'
          schema:
            type: string
      summary: ListPublicationsByUserId lists a user's publications.
      tags:
      - publications
...
swagger: "2.0"